Separate test file creation from benchmarks
This should make both usage and changes to the benchmark much simpler.
Without this commit the benchmark would have too many command line
parameters to easily use, manage and implement. In addition having
the file creation done at the same time increases the risk of failure
and how the benchmarks are implemented the debug prints cannot be used
during benchmarks where as they can be used now with file creation.